requirements.txt
discord.py==1.5.1
PyMySQL==1.0.2
discord==1.0.1
python-valve
pymysql
steamid